+++ server.js
@@ -16,3 +16,4 @@ exports.etag = require('./etag.js');
 exports.router = require('./router.js');
 exports.cache = require('./cache.js');
 exports.cors = require('./cors.js');
+exports.cookie = require('./cookie.js');